Ajuste de interface

+++ src/main/java/com/conversor/domain/Cotacao.java
@@ -9,20 +9,20 @@ public class Cotacao {
     String code;
     String codein;
     String name;
-    float high;
-    float low;
-    float varBid;
-    float pctChange;
-    float bid;
-    float ask;
+    double high;
+    double low;
+    double varBid;
+    double pctChange;
+    double bid;
+    double ask;
     int timestamp;
     String create_date;
 
     public Cotacao() {
     }
 
-    public Cotacao(String code, String codein, String name, float high, float low, float varBid, float pctChange,
-            float bid, float ask, int timestamp, String create_date) {
+    public Cotacao(String code, String codein, String name, double high, double low, double varBid, double pctChange,
+            double bid, double ask, int timestamp, String create_date) {
         this.code = code;
         this.codein = codein;
         this.name = name;
